requirements.txt
numpy
requests
pyalsaaudio
